+++ bin/update.sh
@@ -14,10 +14,14 @@ skips=${SKIP}
 ## Constants
 output_name_updated_count="updated-count"
 output_name_updated_tools="updated-tools"
+output_name_updated_old_versions="updated-old-versions"
+output_name_updated_new_versions="updated-new-versions"
 
 ## State
 remaining_capacity=${max_capacity}
 updated_tools=""
+updated_old_versions=""
+updated_new_versions=""
 
 # --- Import ----------------------------------------------------------------- #
 
@@ -31,6 +35,8 @@ source "${bin}/../lib/actions.sh"
 debug "initializing outputs to their default value"
 set_output "${output_name_updated_count}" "0"
 set_output "${output_name_updated_tools}" "${updated_tools}"
+set_output "${output_name_updated_old_versions}" "${updated_old_versions}"
+set_output "${output_name_updated_new_versions}" "${updated_new_versions}"
 
 debug "checking if .tool-versions file exists"
 if [[ ! -f ".tool-versions" ]]; then
@@ -130,6 +136,22 @@ while read -r line; do
 				updated_tools="${updated_tools},${tool}"
 			fi
 			set_output "${output_name_updated_tools}" "${updated_tools}"
+
+			debug "overriding '${output_name_updated_old_versions}' output with new value"
+			if [ -z "${updated_old_versions}" ]; then
+				updated_old_versions="${current_version}"
+			else
+				updated_old_versions="${updated_old_versions},${current_version}"
+			fi
+			set_output "${output_name_updated_old_versions}" "${updated_old_versions}"
+
+			debug "overriding '${output_name_updated_new_versions}' output with new value"
+			if [ -z "${updated_new_versions}" ]; then
+				updated_new_versions="${latest_version}"
+			else
+				updated_new_versions="${updated_new_versions},${latest_version}"
+			fi
+			set_output "${output_name_updated_new_versions}" "${updated_new_versions}"
 
 			if [ "${remaining_capacity}" -eq 0 ]; then
 				info "finished updating after ${max_capacity} update(s)"