+++ CHANGELOG.md
@@ -12,6 +12,20 @@ Versioning].
 
 ### `tool-versions-update-action`
 
+- _No changes yet._
+
+### `tool-versions-update-action/commit`
+
+- _No changes yet._
+
+### `tool-versions-update-action/pr`
+
+- _No changes yet._
+
+## [0.3.12] - 2023-12-18
+
+### `tool-versions-update-action`
+
 - Add output `updated-new-versions`.
 - Add output `updated-old-versions`.